Create cluster analysis script
Genotype frequency is stored in SQL database along with the superpopulation code SIB AFR, AMR, EAS, EUR, SAS. The user should be able to select which populations or superpopulations (i.e. grouped at continental level) to include in this analysis. If multiple populations are selected in requirement #3, a matrix of pairwise population genetic differentiation should be produced, alongside a suitable visual representation of it. The user should also be able to download a text file with the results obtained. 
This script defines a basic Flask application with three routes:

'/': Displays the main page where users can select the populations or superpopulations.
'/analyze': Handles the form submission from the main page, retrieves genotype frequencies from the database, calculates pairwise population genetic differentiation, visualizes it, and returns the results to the user.
'/download': Allows users to download the visualization of the pairwise population genetic differentiation as a PNG file.
You'll need to create HTML templates (index.html and result.html) to render the main page and display the results, respectively. Additionally, you need to create a SQLite database with the appropriate schema (GenotypeFrequency) and populate it with genotype frequencies data.